// frontend/src/app/core/generated/circabc/model/repositoryConfiguration.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RepositoryConfiguration { 
    name?: string;
    registrationDate?: string;
}